\name{fars}
\docType{data}
\alias{fars}
\title{Fatal Analysis Reporting System}
\description{
  Highway fatality data from the National Highway Safety and Traffic Administration.
}
\usage{fars}
\format{A vector containing over 5000 observations.}
\source{National Highway Traffic Safety Administration}
\keyword{datasets}
